Dissolve AssetManagerDecorator
